@@ -1,15 +1,20 @@
 import React from "react";
-import ReactDOM from "react-dom";
 import "./index.css";
 import App from "./App";
 import { makeServer } from "./server";
+import { BrowserRouter as Router } from "react-router-dom";
+
+import { createRoot } from "react-dom/client";
+
 
 // Call make Server
 makeServer();
 
-ReactDOM.render(
-  <React.StrictMode>
+const container = document.getElementById("root");
+const root = createRoot(container);
+root.render(
+  <Router>
     <App />
-  </React.StrictMode>,
-  document.getElementById("root")
+    </Router>
+
 );
